// src/app/types/type.ts
export interface clothes{
  id?: number;
  product: string;
  price: number;
  stock: number;
  images: string;
}

export interface question{
  id?: number;
  name: string | null;
  email: string | null;
  question: string | null;
}

export interface question{
  id?: number;
  name: string | null;
  email: string | null;
  question: string | null;
}

export interface user{
  id?: string | null;
  name: string | null;
  email: string | null;
  password: string | null;
}